+++ Network/Telegram/API/Bot/Object/Member.hs
@@ -1,8 +1,12 @@
-module Network.Telegram.API.Bot.Object.Member (Member (..)) where
+module Network.Telegram.API.Bot.Object.Member (Member (..), module Exports) where
 
-import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
+import Network.Telegram.API.Bot.Object.Member.Powers as Exports
+import Network.Telegram.API.Bot.Object.Member.Restrictions as Exports
+
+import "aeson" Data.Aeson (FromJSON (parseJSON), Value (Object), withObject, (.:))
 import "base" Control.Applicative (Applicative ((<*>)))
 import "base" Control.Monad (Monad ((>>=)), fail)
+import "base" Data.Bool (Bool)
 import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
@@ -13,9 +17,9 @@ import Network.Telegram.API.Bot.Object.Sender (Sender)
 
 data Member
 	= Creator Sender
-	| Administrator Sender
+	| Administrator Sender Bool Powers
 	| Member Sender
-	| Restricted Sender POSIXTime
+	| Restricted Sender Restrictions POSIXTime
 	| Left Sender
 	| Kicked Sender POSIXTime
 	deriving Show
@@ -23,9 +27,9 @@ data Member
 instance FromJSON Member where
 	parseJSON = withObject "Member" $ \v -> v .: "status" >>= \case
 		("creator" :: Text) -> Creator <$> v .: "user"
-		("administrator" :: Text) -> Administrator <$> v .: "user"
+		("administrator" :: Text) -> Administrator <$> v .: "user" <*> v .: "can_be_edited" <*> parseJSON (Object v)
 		("member" :: Text) -> Member <$> v .: "user"
-		("restricted" :: Text) -> Restricted <$> v .: "user" <*> v .: "until_date"
+		("restricted" :: Text) -> Restricted <$> v .: "user" <*> parseJSON (Object v) <*> v .: "until_date"
 		("left" :: Text) -> Left <$> v .: "user"
 		("kicked" :: Text) -> Kicked <$> v .: "user" <*> v.: "until_date"
 		_ -> fail "Status of chat member is not defined"