+++ Network/API/Telegram/Bot/Object/Update/Callback/Notification.hs
@@ -1,0 +1,3 @@
+module Network.API.Telegram.Bot.Object.Update.Callback.Notification (Notification) where
+
+data Notification