+++ Network/API/Telegram/Bot/Property.hs
@@ -1,0 +1,5 @@
+module Network.API.Telegram.Bot.Property (module Exports) where
+
+import Network.API.Telegram.Bot.Property.Persistable as Exports
+import Network.API.Telegram.Bot.Property.Identifiable as Exports
+import Network.API.Telegram.Bot.Property.Accessible as Exports