+++ Network/Telegram/API/Bot.hs
@@ -1,5 +1,0 @@
-module Network.Telegram.API.Bot (module Exports) where
-
-import Network.Telegram.API.Bot.Property as Exports
-import Network.Telegram.API.Bot.Object as Exports
-import Network.Telegram.API.Bot.Core as Exports