+++ Network/Telegram/API/Bot/Property.hs
@@ -1,5 +1,0 @@
-module Network.Telegram.API.Bot.Property (module Exports) where
-
-import Network.Telegram.API.Bot.Property.Persistable as Exports
-import Network.Telegram.API.Bot.Property.Identifiable as Exports
-import Network.Telegram.API.Bot.Property.Accessible as Exports