+++ CHANGELOG.md
@@ -68,3 +68,4 @@
 * Define `Persistable` instances for `Info` datatype (for direct and reply messages)
 * Add ticks to all constructors of `Capacity` to surround them from two sides in promoted versions
 * Define `Persistable` instance for `Member` (for kicking and unbanning only)
+* Define `Powers` and `Restrictions` datatypes within created `Member` submodule